Migrate the project under trustedfirmware.org governance.
As part of a common agreement between Arm and Google, fork the project
at https://github.com/google/arm-gic and integrate it under the Arm
Firmware Crates umbrella project governed by the TrustedFirmware.org
organization.

As part of this process, change the inbound license to ["MIT" and
"Apache-2.0" and "Developer Certificate of Origin"]. Update the README
file in this regard.

Note that the outbound license does not change, the project is still
dual-licensed under ["MIT" or "Apache-2.0"].

Also embed the contributions guidelines (previously provided as a
separate CONTRIBUTING.MD file) into the README file. This is more
consistent with what other crates at trustedfirmware.org do right now.

Finally, update said contribution guidelines to refer to Gerrit as the
new review system.

Change-Id: Id6fd898b930bde51cff2220b737f06a865ea3a30
Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>