Make distributor and redistributor context sizes configurable
Pack per-interrupt registers into contiguous arrays of
GicDistributorContext and GicRedistributorContext. The size of these
arrays, and thus the number of saved interrupt settings can now be
configured through the const parameters of their structures.
To support this, the patch introduces macros that generate access
functions for the relevant portions of each array. This serves as a
workaround for the current instability of the generic_const_exprs
feature in Rust.

Change-Id: I773f219054ab715dddadb87d37636d548be3f06d
Signed-off-by: Imre Kis <imre.kis@arm.com>